+++ sites/admin-site/src/modules/product-detail/components/index.ts
@@ -1,2 +1,3 @@
 export * from './VariantLineItem';
 export * from './SkuForm';
+export * from './ProductSkuInventoryList';
