+++ sites/admin-site/src/modules/product-detail/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './useProductContext';
 export * from './useVariants';
 export * from './useProductDetail';
 export * from './useProductSkus';
+export * from './useSkuInventory';
