+++ sites/admin-site/src/modules/product-detail/product-detail-service.ts
@@ -4,12 +4,10 @@ import { IProductInfoFormType } from 'modules/products/types';
 import {
   IGetListVariantsResponse,
   IUpsertVariantPayload,
-} from './types.ts/variant';
-import {
-  IGetListSkusResponse,
-  IUpsertSkuFormType,
-} from './types.ts/product-skus';
+} from './types/variant';
+import { IGetListSkusResponse, IUpsertSkuFormType } from './types/product-skus';
 import { ISeoInfo } from 'modules/seo-info/types';
+import { IGetSkuInventoryListResponse } from './types';
 
 export const productDetailService = {
   getProductDetail: async (id: ID) => {
@@ -56,6 +54,14 @@ export const productDetailService = {
   updateSkuSeo: async (productId: ID, sku: ID, seoInfo: ISeoInfo) => {
     return appApi
       .post(`/products/${productId}/skus/${sku}/seo`, seoInfo)
+      .then((res) => res.data);
+  },
+
+  getAllSkuInventory: async (productId: ID) => {
+    return appApi
+      .get<IGetSkuInventoryListResponse>(`/sku-inventory`, {
+        params: { productId },
+      })
       .then((res) => res.data);
   },
 };
