+++ sites/admin-site/src/modules/product-detail/types/index.ts
@@ -1,1 +1,2 @@
 export * from './product-detail';
+export * from './inventory';
